pyproject.toml
[tool.poetry]
name = "maubot"
version = "1.10"
description = "Play Uno with your friends in Telegram chats."
authors = ["Milinuri Nirvalen <pentergust@mail.ru>"]
license = "AGPL-3.0-or-later"
readme = "README.md"


# Dependencies ---------------------------------------------------------

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.3"
aiogram = "^3.18"
tortoise-orm = "^0.24.0"
pydantic-settings = "^2.8.0"


# Mau backend dependencies ---------------------------------------------

[tool.poetry.group.web.dependencies]
fastapi = "^0.115.8"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"


# Development dependencies ---------------------------------------------

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
icecream = "^2.1.3"
telethon = "^1.39.0"
mypy = "^1.15.0"


# Ruff linter ----------------------------------------------------------

[tool.ruff]
line-length = 80
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "D", "N", "PL", "UP", "ANN", "W", "ASYNC"]
ignore = ["D407", "D107", "D213", "D203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


# Configure build ------------------------------------------------------

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
